% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/strataG-package.R
\docType{data}
\name{dolph.seqs}
\alias{dolph.seqs}
\title{Dolphin mtDNA Sequences}
\format{A list of 126 aligned control region sequences}
\usage{
data(dolph.seqs)
}
\description{
Dolphin mtDNA Sequences
}
\references{
Lowther et al.
}
\keyword{datasets}

