% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/strataG-package.R
\docType{data}
\name{dolph.strata}
\alias{dolph.strata}
\title{Dolphin Genetic Stratification and Haplotypes}
\format{A data.frame of 126 samples and 4 variables}
\usage{
data(dolph.strata)
}
\description{
Assignment of samples to one of two stratifications and mtDNA haplotype designations
}
\references{
Lowther et al.
}
\keyword{datasets}

