% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/exptdHet.R
\name{exptdHet}
\alias{exptdHet}
\title{Expected Heterozygosity}
\usage{
exptdHet(g)
}
\arguments{
\item{g}{a \linkS4class{gtypes} object.}
}
\description{
Calculate expected heterozygosity for diploid data.
}
\examples{
data(dolph.msats)
data(dolph.strata)
msats.merge <- merge(dolph.strata[, c("ids", "fine")], dolph.msats,
  all.y = TRUE)
msats.g <- df2gtypes(msats.merge, ploidy = 2)
exptdHet(msats.g)
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
\seealso{
\code{\link{obsvdHet}}
}

