% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fastsimcoal.R
\name{fastsimcoal}
\alias{fastsimcoal}
\title{Run fastsimcoal}
\usage{
fastsimcoal(num.pops, Ne, sample.size = NULL, sample.time = NULL,
  growth.rate = NULL, mig.rates = NULL, hist.ev = NULL, num.chrom = 1,
  locus.params = NULL, label = "strataG.fastsimcoal", num.sims = 1,
  inf.site.model = TRUE, quiet = TRUE, delete.files = TRUE)
}
\arguments{
\item{num.pops}{number of populations.}

\item{Ne}{effective population size.}

\item{sample.size}{number of samples to take.}

\item{sample.time}{time to draw samples.}

\item{growth.rate}{growth rate of populations.}

\item{mig.rates}{list of migration matrices.}

\item{hist.ev}{historical events.}

\item{num.chrom}{number of chromosomes.}

\item{locus.params}{locus parameters.}

\item{label}{character string to label files with.}

\item{num.sims}{number of simulations to run.}

\item{inf.site.model}{logical. Infinite site model?}

\item{quiet}{logical. Run quietly?}

\item{delete.files}{logical. Delete files when done?}
}
\value{
A list of \code{\link{gtypes}} objects for each simulated dataset.
}
\description{
Run a fastsimcoal simualtion and load results into a
  \linkS4class{gtypes} object.
}
\note{
Assumes that the program \code{fastsimcoal} is properly installed and
  available on the command line. On PC's, this requires having it in a
  folder in the PATH environmental variable. On Macs, the executable
  should be installed in a folder like \code{/usr/local/bin}
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
\references{
Excoffier, L. and Foll, M (2011) fastsimcoal: a continuous-time
  coalescent simulator of genomic diversity under arbitrarily complex
  evolutionary scenarios Bioinformatics 27: 1332-1334.\cr
  \url{http://cmpg.unibe.ch/software/fastsimcoal2/}
}

