% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fstToNm.R
\name{fstToNm}
\alias{fstToNm}
\title{Nm from Fst}
\usage{
fstToNm(fst, diploid = TRUE)
}
\arguments{
\item{fst}{estimate of Fst between populations.}

\item{diploid}{logical. Is Fst from diploid (TRUE) or haploid (FALSE) data?}
}
\description{
Calculate Nm for a given value of Fst.
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}

