% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/genind2gtypes.R
\name{genind2gtypes}
\alias{genind2gtypes}
\alias{gtypes2genind}
\title{Convert Between \code{gtypes} And \code{genind} objects.}
\usage{
genind2gtypes(x)

gtypes2genind(x, type = c("codom", "PA"))
}
\arguments{
\item{x}{either a \linkS4class{gtypes} or \linkS4class{genind} object
to convert from.}

\item{type}{a character string indicating the type of marker for
  \linkS4class{genind} objects: 'codom' stands for 'codominant'
  (e.g. microstallites, allozymes); 'PA' stands for 'presence/absence'
  markers (e.g. AFLP, RAPD).}
}
\description{
Convert a \code{gtypes} object to a \code{genind} object
  and vice-versa.
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}

