% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/genotype.R
\name{genotype}
\alias{genotype}
\title{Genotype}
\usage{
genotype(ids, loci, g)
}
\arguments{
\item{ids}{character vector of individual ids.}

\item{loci}{character vector of loci.}

\item{g}{a \linkS4class{gtypes} object.}
}
\value{
a data.frame of the genotypes requested.
}
\description{
Get genotype of a specified set of individuals and loci
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}

