% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Hstats.R
\name{Hstats}
\alias{Hstats}
\title{Heterozygosity estimates: Ho, Hs, and Ht}
\usage{
Hstats(g)
}
\arguments{
\item{g}{a \code{gtypes} object}
}
\value{
matrix of Ho, Hs, and Ht (rows) for each locus (columns)
}
\description{
Calculate overall heterozygosity (Ho), expected
  heterozygosity in subpopulations (Hs), and expected total
  heterozygosity (Ht)
}
\author{
Eric Archer <eric.archer@noaa.gov>
}
\references{
Nei, M. and R.K. Chesser. 1983. Estimation of fixation
  indices and gene diversities. Ann. Hum. Genet. 47:253-259.
}

