% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/idRows.R
\name{idRows}
\alias{idRows}
\title{\code{gtypes} ID Rows}
\usage{
idRows(ids, rowNames)
}
\arguments{
\item{ids}{character vector of sample ids.}

\item{rowNames}{a vector of rownames from the \code{@loci} slot of a
  \linkS4class{gtypes} object.}
}
\value{
a vector of row numbers in \code{g@loci} where data for \code{ids}
  are stored.
}
\description{
Return row numbers of \code{@loci} slot for a set of
  sample ids.
}
\examples{
data(dolph.msats)
data(dolph.strata)
msats.merge <- merge(dolph.strata[, c("ids", "fine")], dolph.msats, all.y = TRUE)
msats <- df2gtypes(msats.merge, ploidy = 2)

ran.ids <- sample(indNames(msats), 5)
idRows(ran.ids, rownames(loci(msats)))
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}

