% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/locusNames.R
\name{locusNames}
\alias{locusNames}
\title{Locus Names}
\usage{
locusNames(loci, ploidy, default.name = "Locus")
}
\arguments{
\item{loci}{a vector or matrix of alleles. If a matrix, each allele is in
consecutive columns of the matrix.}

\item{ploidy}{integer specifying the ploidy}

\item{default.name}{name to provide for loci if no column names are provided
  in \code{loci}.}
}
\value{
A vector of locus names.
}
\description{
Parse locus names from a matrix where alleles are in
  consecutive columns.
}
\note{
If \code{loci} has column names, then columns of each locus must be
  named with the same roots. For example, diploid locus 'ABCD' would have
  two columns named something like 'ABCD.1' and 'ABCD.2', or
  'ABCD_A' and 'ABCD_B'. \cr
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}

