% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/strataG-package.R
\docType{data}
\name{msats.g}
\alias{msats.g}
\title{Dolphin Microsatellite gtypes Object}
\format{A \linkS4class{gtypes} object of 126 samples and 15 loci}
\usage{
data(msats.g)
}
\description{
Genotypes for 15 microsatellite loci
}
\references{
Lowther-Thieleking J.L., F.I. Archer, A.R. Lang, and
  D.W. Weller. 2015. Genetic variation of coastal and offshore bottlenose
  dolphins, Tursiops truncatus, in the eastern North Pacific Ocean.
  Marine Mammal Science 31:1-20
}
\keyword{datasets}

