% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nucleotideDiversity.R
\name{nucleotideDiversity}
\alias{nucleotideDiversity}
\title{Nucleotide Diversity}
\usage{
nucleotideDiversity(x, bases = c("a", "c", "g", "t"))
}
\arguments{
\item{x}{a set of sequences or a \linkS4class{gtypes} object with sequences.}

\item{bases}{nucleotides to consider when calculating diversity.}
}
\value{
Nucleotide diversity by site.
}
\description{
Calculate nucleotide diversity for set of haplotypes.
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}

