% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/obsvdHet.R
\name{obsvdHet}
\alias{obsvdHet}
\title{Observed Heterozygosity}
\usage{
obsvdHet(g)
}
\arguments{
\item{g}{a \linkS4class{gtypes} object.}
}
\description{
Calculate observed heterozygosity for diploid data.
}
\examples{
data(dolph.msats)
data(dolph.strata)
msats.merge <- merge(dolph.strata[, c("ids", "fine")], dolph.msats,
  all.y = TRUE)
msats.g <- df2gtypes(msats.merge, ploidy = 2)
obsvdHet(msats.g)
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
\seealso{
\code{\link{exptdHet}}
}

