% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pctUniqueAlleles.R
\name{pctUniqueAlleles}
\alias{pctUniqueAlleles}
\title{Percent Unique Alleles}
\usage{
pctUniqueAlleles(g)
}
\arguments{
\item{g}{a \linkS4class{gtypes} object.}
}
\value{
a vector of the number of unique (occuring only in one individual)
  alleles for each locus.
}
\description{
Calculate the percent of alleles that are unique.
}
\examples{
data(dolph.strata)
data(dolph.seqs)
strata <- dolph.strata$fine
names(strata) <- dolph.strata$ids
dloop <- sequence2gtypes(dolph.seqs, strata, seq.names = "dLoop")
dloop <- labelHaplotypes(dloop)$gtypes

pctUniqueAlleles(dloop)
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}

