% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pctUniqueHaps.R
\name{pctUniqueHaps}
\alias{pctUniqueHaps}
\title{Percent Unique Haplotypes}
\usage{
pctUniqueHaps(g)
}
\arguments{
\item{g}{a \linkS4class{gtypes} object.}
}
\description{
Calculate the percent of haplotypes that are unique.
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}

