% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/permuteStrata.R
\name{permuteStrata}
\alias{permuteStrata}
\title{Permute strata}
\usage{
permuteStrata(g)
}
\arguments{
\item{g}{a \linkS4class{gtypes} object.}
}
\description{
Permute the strata slot within a \linkS4class{gtypes} object.
}
\examples{
data(dolph.msats)
data(dolph.strata)
msats.merge <- merge(dolph.strata[, c("ids", "fine")], dolph.msats,
  all.y = TRUE, description = date())
msats <- df2gtypes(msats.merge, ploidy = 2)
summary(msats)
ran.msats <- permuteStrata(msats)
summary(ran.msats)
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}

