% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/popStructTest.R
\name{popStructTest}
\alias{overallTest}
\alias{pairwiseTest}
\alias{popStructTest}
\alias{statList}
\title{Population Differentiation Tests}
\usage{
popStructTest(g, nrep = 100, stats = "all", type = "both",
  keep.null = FALSE, quietly = FALSE, num.cores = 1,
  write.output = FALSE, ...)

overallTest(g, nrep = 100, stat.list = statList("all"), keep.null = FALSE,
  quietly = FALSE, num.cores = 1, ...)

pairwiseTest(g, nrep = 100, stat.list = statList("all"),
  keep.null = FALSE, quietly = TRUE, num.cores = 1, ...)

statList(stats = "all")
}
\arguments{
\item{g}{a \code{\link{gtypes}} object.}

\item{nrep}{number specifying number of permutation replicates to use for
permutation test.}

\item{stats}{a character vector specifying which anlayses to conduct.
Possible choices are: "phist", "fst", "fst.prime", "gst", "gst.prime",
"gst.dbl.prime", "d", or "chi2", or "all".}

\item{type}{character determining type of test to conduct. Can be "overall",
"pairwise", or "both". If "pairwise" or "both" are chosen and there are
only two strata, then only an overall test will be conducted.}

\item{keep.null}{logical. Keep the null distribution from the
permutation test?}

\item{quietly}{logical. Print progress and results?}

\item{num.cores}{number of CPU cores to use. Value is passed to
\code{\link[parallel]{mclapply}}.}

\item{write.output}{logical. Write a .csv file with results?}

\item{...}{other parameters to be passed to population
  differentiation functions.}

\item{stat.list}{a list of popStructStat functions.
Generated by \code{statList}.}
}
\value{
\describe{
 \item{overall}{a list containing:
   \tabular{ll}{
     \code{strata.freq} \tab a vector of the sample sizes for each stratum.\cr
     \code{result} \tab a matrix with the statistic estimate and p-value
       for each statistic.\cr
     \code{null.dist} \tab a matrix with the null distributions for
       each statistic.\cr
   }}
 \item{pairwise}{a list containing:
   \tabular{ll}{
     \code{result} \tab a data.frame with the result of each pairwise
       comparison on a line.\cr
     \code{pair.mat} \tab a list with a pairwise matrix for each statistic.
       Values in lower left are p-values, upper right is statistic estimate.\cr
     \code{null.dist} \tab a matrix with the null distributions for
       each statistic.\cr
   }}
}
}
\description{
Conduct overall and/or pairwise tests of
  population differentiation.
}
\examples{
data(dolph.msats)
data(dolph.strata)
msats.merge <- merge(dolph.strata[, c("ids", "fine")], dolph.msats,
  all.y = TRUE)
msats <- df2gtypes(msats.merge, ploidy = 2)

# Conduct an overall Chi-squared test
ovl <- overallTest(msats, nrep = 10,
  stat.list = statList("chi2"), quietly = FALSE
)

# Conduct a pairwise test for Gst
pws <- pairwiseTest(msats, nrep = 10,
  stat.list = list(statGst), quietly = FALSE
)

# Conduct both overall and pairwise tests for Fst and F'st
full <- popStructTest(msats, nrep = 10, stats = "fst")
print(full$overall)
print(full$pairwise)
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}

