% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/qaqcSummary.R
\name{qaqcSummary}
\alias{qaqcSummary}
\title{QA/QC Summaries}
\usage{
qaqcSummary(g, label = NULL, ...)
}
\arguments{
\item{g}{a \linkS4class{gtypes} object.}

\item{label}{optional label for output folder and prefix for files.}

\item{...}{optional arguments to pass on to summary functions.}
}
\description{
QA/QC Summaries
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}

