% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sharedLoci.R
\name{sharedLoci}
\alias{propSharedIds}
\alias{propSharedLoci}
\alias{sharedAlleles}
\alias{sharedLoci}
\title{Shared Loci}
\usage{
propSharedLoci(g, type = "strata", num.cores = 1)

sharedAlleles(g, smry = "num")

propSharedIds(ids, g)
}
\arguments{
\item{g}{a \code{\link{gtypes}} object.}

\item{type}{a character vector determining type of pairwise comparsion. Can
be "ids" for individuals or "strata" for strata.}

\item{num.cores}{number of CPU cores to use. Value is passed to
  \code{\link[parallel]{mclapply}}.}

\item{smry}{a character vector determining type of summary for
\code{sharedAlleles}. "which" returns the names of the alleles shared.
"num" returns the number of alleles shared.}

\item{ids}{character vector of two sample ids to compare.}
}
\value{
matrix summary of pairwise shared loci.
}
\description{
Calculate proportion of alleles and number of loci shared
  between pairs of individuals or strata.
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}

