% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/statChi2.R
\name{statChi2}
\alias{statChi2}
\title{CHI-squared Analysis of Population Structure}
\usage{
statChi2(g)
}
\arguments{
\item{g}{a \code{gtypes} object.}
}
\value{
a list with the CHI-squared estimate.
}
\description{
CHI-squared Analysis of Population Structure
}
\examples{
data(dolph.strata)
data(dolph.msats)
#msat.merge <- merge(dolph.strata, dolph.msats, by = "ids", all.y = TRUE, sort = FALSE)
#msats <- gtypes(msat.merge, id.col = 1, strata.col = 3, locus.col = 5, description = "msats")

#stat.chi2(msats)
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}

