% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/strataPairs.R
\name{strataPairs}
\alias{strataPairs}
\title{Strata Pairs}
\usage{
strataPairs(g)
}
\arguments{
\item{g}{a \linkS4class{gtypes} object.}
}
\description{
Make a matrix of all pairs of strata.
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}

