% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/strataSplit.R
\name{strataSplit}
\alias{strataSplit}
\title{Split Strata}
\usage{
strataSplit(g, strata = NULL)
}
\arguments{
\item{g}{a \linkS4class{gtypes} object.}

\item{strata}{a character vector giving a subset of strata to select.
  If \code{NULL} then a list with all strata is created.}
}
\value{
A named list where each element is a \code{gtypes} object
  for a single stratum in \code{g}.
}
\description{
Return a list of gtypes for each strata
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}

