% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/structureParseQmat.R
\name{structureParseQmat}
\alias{structureParseQmat}
\title{Parse Q-Matrix Text}
\usage{
structureParseQmat(q.mat.txt, pops)
}
\arguments{
\item{q.mat.txt}{character vector of Q matrix from STRUCTURE output file.}

\item{pops}{vector of population labels to be used in place of numbers in
  STRUCTURE file.}
}
\description{
Parse Q-Matrix Text
}
\details{
Function is used internally by \code{\link{structure}} and
  \code{\link{clumpp}} to parse output files.
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}

