% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/subset.gtypes.R
\docType{methods}
\name{subset,gtypes-method}
\alias{subset,gtypes}
\alias{subset,gtypes-method}
\title{Subset gtypes}
\usage{
\S4method{subset}{gtypes}(x, strata = NULL, ids = NULL, loci = NULL,
  remove.sequences = FALSE, ...)
}
\arguments{
\item{x}{a \linkS4class{gtypes} object.}

\item{strata}{a character vector designating individuals of which strata
to select.}

\item{ids}{a character vector designating which individuals to select.}

\item{loci}{a character vector designating which loci to select.}

\item{remove.sequences}{logical. If \code{TRUE} any sequences not referenced
in selected samples will not be in the returned object.}

\item{...}{optional arguments passed through generic (ignored).}
}
\value{
a \linkS4class{gtypes} object containing only individuals and loci
  that match the above criteria. If no individuals match (e.g., all ids are
  in non-selected strata), the return is \code{NULL}.
}
\description{
Return subsets of a gtypes object based on specified
  strata, individuals, or loci.
}
\examples{
data(dolph.msats)
data(dolph.strata)
msats.merge <- merge(dolph.strata[, c("ids", "broad")], dolph.msats, all.y = TRUE)
msats <- df2gtypes(msats.merge, ploidy = 2)

# just Coastal genotypes
msats.coastal <- subset(msats, strata = "Coastal")
msats.coastal

# pick 10 random offshore samples and the first two loci
msats.offshore <- subset(msats, strata = "Offshore")
ran.ids <- sample(indNames(msats.offshore), 10)
loc2 <- locNames(msats.offshore)[1:2]
msats.offshore <- subset(msats.offshore, ids = ran.ids, loci = loc2)
msats.offshore
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}

