% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/summarizeLoci.R
\name{summarizeLoci}
\alias{summarizeLoci}
\title{Locus Summaries}
\usage{
summarizeLoci(g, by.strata = FALSE, ...)
}
\arguments{
\item{g}{a \linkS4class{gtypes} object.}

\item{by.strata}{logical. If \code{TRUE}, return a list of summary matrices
for each stratum.}

\item{...}{arguments to be passed on to summary functions.}
}
\value{
A matrix with rows for each locus and columns containing summary
  statistics.
}
\description{
Compile standard by-locus summaries.
}
\examples{
data(dolph.msats)
data(dolph.strata)
msats.merge <- merge(dolph.strata[, c("ids", "fine")], dolph.msats, all.y = TRUE)
msats <- df2gtypes(msats.merge, ploidy = 2)

summarizeLoci(msats)
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}

