% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/trimNs.R
\name{trimNs}
\alias{trimNs}
\title{Trim N's From Sequences}
\usage{
trimNs(x)
}
\arguments{
\item{x}{a list of DNA sequences.}
}
\value{
a list of DNA sequences.
}
\description{
Removes N's from beginning and end of sequences.
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}

