% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/validIupacCodes.R
\name{validIupacCodes}
\alias{validIupacCodes}
\title{Valid IUPAC Codes}
\usage{
validIupacCodes(bases)
}
\arguments{
\item{bases}{character vector of nucleotides or IUPAC codes to be checked.}
}
\value{
character vector of valid IUPAC codes for \code{bases}.
}
\description{
Get all possible valid IUPAC codes.
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}

