% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/variableSites.R
\name{variableSites}
\alias{variableSites}
\title{Variable Sites}
\usage{
variableSites(x, bases = c("a", "c", "g", "t", "-"))
}
\arguments{
\item{x}{a \code{\link{gtypes}} object with sequences, a list of sequences, or a consensus sequence.
Sequences must be aligned.}

\item{bases}{character vector of bases to consider.}
}
\value{
A list with: \tabular{ll}{
  \code{site} \tab a list of sequences composed of variable sites. \cr
  \code{site.freqs} \tab a matrix of base pair frequencies by site. \cr
}
}
\description{
Identify variable sites among sequences.
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
\seealso{
\code{\link{fixedSites}}
}

