% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/write.gtypes.R
\name{writeGtypes}
\alias{writeGtypes}
\title{Write gtypes}
\usage{
writeGtypes(g, label = NULL, folder = NULL, as.frequency = FALSE)
}
\arguments{
\item{g}{a \linkS4class{gtypes} object.}

\item{label}{label for filename(s). Default is the gtypes description
if present.}

\item{folder}{folder where file(s) should be written to. If NULL, files
are written to current working directory.}

\item{as.frequency}{logical indicating if haploid data should be output
  as frequency tables.}
}
\description{
Write a \code{\link{gtypes}} object to file(s).
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}

