% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/alleleFreqs.R
\name{alleleFreqs}
\alias{alleleFreqs}
\title{Allele Frequencies}
\usage{
alleleFreqs(g, by.strata = FALSE)
}
\arguments{
\item{g}{a \linkS4class{gtypes} object.}

\item{by.strata}{logical. If \code{TRUE} every element in the return list is
  a three dimensional array where the third dimension contains frequencies
  and proportions for each stratum.}
}
\value{
A list of allele frequencies for each locus. Each element is a
  matrix or array with frequencies by count (\code{freq}) and
  proportion (\code{prop}) of each allele.
}
\description{
Calculate allele frequencies for each locus.
}
\examples{
data(dolph.msats)
data(dolph.strata)
msats.merge <- merge(dolph.strata[, c("ids", "fine")], dolph.msats, all.y = TRUE)
msats <- df2gtypes(msats.merge, ploidy = 2)

f <- alleleFreqs(msats)
f[[1]]

f.pop <- alleleFreqs(msats, TRUE)
f.pop[[1]]
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}

