% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/alleleFreqs.R
\name{alleleFreqs}
\alias{alleleFreqs}
\title{Allele Frequencies}
\usage{
alleleFreqs(g, by.strata = FALSE)
}
\arguments{
\item{g}{a \linkS4class{gtypes} object.}

\item{by.strata}{if \code{TRUE} every element in the return list is a three
  dimensional array where the third dimension contains frequencies and
  proportions for each stratum.}
}
\value{
A list of allele frequencies for each locus. Each element is a
  data.frame with frequencies by count (freq) and proportion (prop) of
  each allele.
}
\description{
Calculate allele frequencies for each locus.
}
\examples{
data(dolph.msats)
data(dolph.strata)
strata.schemes <- dolph.strata[, c("broad", "fine")]
rownames(strata.schemes) <- dolph.strata$id
msats <- new("gtypes", gen.data = dolph.msats[, -1], ploidy = 2,
             ind.names = dolph.msats[, 1], schemes = strata.schemes,
             strata = "fine")

f <- alleleFreqs(msats)
f[[1]]

f.pop <- alleleFreqs(msats, TRUE)
f.pop[[1]]
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}

