% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/allelicRichness.R
\name{allelicRichness}
\alias{allelicRichness}
\title{Allelic Richness}
\usage{
allelicRichness(g)
}
\arguments{
\item{g}{a \linkS4class{gtypes} object.}
}
\description{
Calculate allelic richness for each locus.
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}

