% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gtypes.accessors.R
\docType{class}
\name{gtypes.accessors}
\alias{description}
\alias{description,gtypes}
\alias{description,gtypes-method}
\alias{gtypes.accessors}
\alias{indNames,gtypes}
\alias{indNames,gtypes-method}
\alias{locNames,gtypes}
\alias{locNames,gtypes-method}
\alias{nInd,gtypes}
\alias{nInd,gtypes-method}
\alias{nLoc,gtypes}
\alias{nLoc,gtypes-method}
\alias{nStrata}
\alias{nStrata,gtypes}
\alias{nStrata,gtypes-method}
\alias{other,gtypes}
\alias{other,gtypes-method}
\alias{ploidy,gtypes}
\alias{ploidy,gtypes-method}
\alias{schemes}
\alias{schemes,gtypes}
\alias{schemes,gtypes-method}
\alias{schemes<-}
\alias{schemes<-,gtypes}
\alias{schemes<-,gtypes-method}
\alias{seqNames}
\alias{seqNames,gtypes}
\alias{seqNames,gtypes-method}
\alias{sequences}
\alias{sequences,gtypes}
\alias{sequences,gtypes-method}
\alias{strata}
\alias{strata,gtypes}
\alias{strata,gtypes-method}
\alias{strata<-}
\alias{strata<-,gtypes}
\alias{strata<-,gtypes-method}
\alias{strataNames}
\alias{strataNames,gtypes}
\alias{strataNames,gtypes-method}
\title{gtypes Accessors}
\usage{
\S4method{nInd}{gtypes}(x, ...)

\S4method{nLoc}{gtypes}(x, ...)

nStrata(x, ...)

\S4method{nStrata}{gtypes}(x, ...)

\S4method{indNames}{gtypes}(x, ...)

\S4method{locNames}{gtypes}(x, ...)

strataNames(x, ...)

\S4method{strataNames}{gtypes}(x, ...)

\S4method{ploidy}{gtypes}(x, ...)

\S4method{other}{gtypes}(x, ...)

strata(x, ...)

\S4method{strata}{gtypes}(x, ...)

strata(x) <- value

\S4method{strata}{gtypes}(x) <- value

schemes(x, ...)

\S4method{schemes}{gtypes}(x, ...)

schemes(x) <- value

\S4method{schemes}{gtypes}(x) <- value

seqNames(x, ...)

\S4method{seqNames}{gtypes}(x, ...)

sequences(x, ...)

\S4method{sequences}{gtypes}(x, ...)

description(x, ...)

\S4method{description}{gtypes}(x, ...)
}
\arguments{
\item{x}{a \linkS4class{gtypes} object.}

\item{...}{other arguments passed from generics (ignored).}

\item{value}{value being assigned by accessor.}
}
\description{
Accessors for slots in \linkS4class{gtypes} objects.
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}

