% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/idRows.R
\name{idRows}
\alias{idRows}
\title{ID Rows}
\usage{
idRows(ids, g)
}
\arguments{
\item{ids}{character vector of sample ids.}

\item{g}{a \linkS4class{gtypes} object.}
}
\value{
a vector of row numbers in \code{g@loci} where data for \code{ids}
  are stored.
}
\description{
Return row numbers of \code{@loci} slot for a set of
  sample ids.
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}

