% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/subset.gtypes.R
\docType{methods}
\name{subset,gtypes-method}
\alias{subset,gtypes}
\alias{subset,gtypes-method}
\title{Subset gtypes}
\usage{
\S4method{subset}{gtypes}(x, strata = NULL, ids = NULL, loci = NULL,
  remove.sequences = FALSE, ...)
}
\arguments{
\item{x}{a \linkS4class{gtypes} object.}

\item{strata}{a character vector designating individuals of which strata
to select.}

\item{ids}{a character vector designating which individuals to select.}

\item{loci}{a character vector designating which loci to select.}

\item{remove.sequences}{logical. If \code{TRUE} any sequences not referenced
in selected samples will not be in the returned object.}

\item{...}{optional arguments passed through generic (ignored).}
}
\value{
a \linkS4class{gtypes} object containing only individuals and loci
  that match the above criteria. If no individuals match (e.g., all ids are
  in non-selected strata), the return is \code{NULL}.
}
\description{
Return subsets of a gtypes object based on specified
  strata, individuals, or loci.
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}

